Make SmartPlaylist child of Playlist
